backend/requirements_linux.txt
Django==4.1.2
django-cors-headers==3.13.0
django-phonenumber-field==7.0.0
phonenumbers==8.12.57
Pillow==9.2.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0
drf-spectacular==0.24.2
django-silk==5.0.2
PyJWT==2.6.0
redis==4.4.0
pre-commit==2.21.0
channels==4.0.0
daphne==4.0.0
channels-redis==4.0.0
